+++ admin/controller/payment/duitku_va_ritel.php
@@ -33,6 +33,7 @@ class ControllerPaymentDuitkuVARitel extends Controller {
       'text_edit',
       'entry_merchant',
       'entry_api_key',          
+      'entry_expired_period',          
       'entry_test',
       'entry_total',
       'entry_order_status',
@@ -87,6 +88,8 @@ class ControllerPaymentDuitkuVARitel extends Controller {
       'duitku_va_ritel_merchant',
       'duitku_va_ritel_environment',
       'duitku_va_ritel_api_key',
+      'duitku_va_ritel_expired',
+      'duitku_va_ritel_expired',
       'duitku_va_ritel_endpoint',      
       'duitku_va_ritel_debug',
       'duitku_va_ritel_total',
@@ -148,19 +151,23 @@ class ControllerPaymentDuitkuVARitel extends Controller {
     }
         
 
-      // check for empty values
-      if (!$this->request->post['duitku_va_ritel_api_key']) {
-        $this->error['client_key_v2'] = $this->language->get('error_client_key');
-      }
+	// check for empty values
+	if (!$this->request->post['duitku_va_ritel_api_key']) {
+		$this->error['client_key_v2'] = $this->language->get('error_client_key');
+	}
 
-      if (!$this->request->post['duitku_va_ritel_merchant']) {
-        $this->error['server_key_v2'] = $this->language->get('error_server_key');
-      }        
-      
+	if (!$this->request->post['duitku_va_ritel_merchant']) {
+		$this->error['server_key_v2'] = $this->language->get('error_server_key');
+	}        
 
-     if (!$this->request->post['duitku_va_ritel_endpoint']) {
-        $this->error['endpoint'] = $this->language->get('error_endpoint');
-      }        
+	if (!$this->request->post['duitku_va_ritel_expired'] OR $this->request->post['duitku_va_ritel_expired'] > 1440 ) {
+		$this->error['expired_period'] = $this->language->get('error_expired_period');
+	}        
+
+
+	if (!$this->request->post['duitku_va_ritel_endpoint']) {
+		$this->error['endpoint'] = $this->language->get('error_endpoint');
+	}        
 
     if (!$this->error) {
       return true;